first working compilation
