+++ config/splat.us.ric.yaml
@@ -41,9 +41,9 @@ segments:
       - [0x19BE0, data]
       - [0x19C70, data]
       - [0x19EE0, data]
-      - [0x1A444, rodata]
+      - [0x1A444, .rodata, 1AC60] #UpdateEntityRichter
       - [0x1A5E0, .rodata, 1AC60]
-      - [0x1A788, .rodata, 1F348]
+      - [0x1A5F0, .rodata, 1CB04]
       - [0x1A7A8, .rodata, 1FCD0]
       - [0x1A7C0, .rodata, 202A8]
       - [0x1A800, .rodata, 20920]
@@ -57,7 +57,7 @@ segments:
       - [0x1AB68, .rodata, 32324]
       - [0x1AC50, rodata]
       - [0x1AC60, c, 1AC60]
-      - [0x1F348, c, 1F348]
+      - [0x1CB04, c, 1CB04]
       - [0x1FCD0, c, 1FCD0]
       - [0x202A8, c, 202A8]
       - [0x20920, c, 20920]
