+++ src/ric/1AC60.c
@@ -360,7 +360,346 @@ void CheckHighJumpInput(void) {
     }
 }
 
-INCLUDE_ASM("ric/nonmatchings/1AC60", UpdateEntityRichter);
+extern s16 D_80154574[];
+extern s16 D_80154594[];
+extern s32 D_8015538C;
+extern AnimationFrame D_80155658[];
+extern AnimationFrame D_801556C4[];
+extern FrameProperty D_80155964;
+
+void UpdateEntityRichter(void) {
+    DamageParam sp10;
+    s32 temp_s0;
+    s32 var_s4;
+    s32 var_fp;
+    s32 var_s7;
+    s16 var_s5;
+    s16 var_s6;
+    s32 i;
+    bool condition;
+    f32* playerY;
+    
+
+    PlayerDraw* playerDraw = g_PlayerDraw;
+
+    g_CurrentEntity = &PLAYER;
+    g_Player.unk4C = 0;
+    g_Player.unk72 = func_80156DE4();
+    FntPrint("pl_head_f:%02x\n", g_Player.unk72);
+    for(i = 0; i < 16; i++) {
+        if (g_Player.D_80072F00[i] == 0) {
+            continue;
+        }
+        switch (i) {
+        case 0:          
+        case 1:          
+        case 3:          
+        case 5:          
+        case 6:          
+        case 7:          
+        case 8:          
+        case 9:          
+        case 10:         
+        case 11:         
+        case 12:         
+        case 14:         
+            break;
+        case 2:          
+            PLAYER.palette = g_Player.unk40;
+            break;
+        case 4:{
+            s32 temp_s0 = (g_GameTimer & 0xF) << 8;
+            playerDraw->r0 = playerDraw->b0 = playerDraw->g0 = (rsin((s16)temp_s0 + 0x000) + 0x1000) / 64 + 0x60;
+            playerDraw->r1 = playerDraw->b1 = playerDraw->g1 = (rsin(temp_s0 + 0x200) + 0x1000) / 64 + 0x60;
+            playerDraw->r3 = playerDraw->b3 = playerDraw->g3 = (rsin(temp_s0 + 0x400) + 0x1000) / 64 + 0x60;
+            playerDraw->r2 = playerDraw->b2 = playerDraw->g2 = (rsin(temp_s0 + 0x600) + 0x1000) / 64 + 0x60;
+            playerDraw->enableColorBlend = 1;
+            break;
+        }
+        case 13:         
+            g_Player.D_80072F00[13] = 4;
+            break;
+        case 15:         
+            func_8015CB58(0, 0);
+            break;
+        }
+        if (--(g_Player.D_80072F00[i]) != 0) {
+            continue;
+        }
+        switch (i) {
+        case 0:
+            break;
+        case 2:          
+            PLAYER.palette = 0x8120;
+            break;
+        case 4:          
+            playerDraw->enableColorBlend = 0;
+            break;
+        case 13:         
+            func_8015CAD4(1, 0x10);
+            break;
+        case 6:          
+            if ((PLAYER.step == 3) && (PLAYER.unk4C != D_80155534)) {
+                func_8015C920(D_80155534);
+                g_Player.unk44 &= 0xFFEF;
+            }
+            break;
+        case 15:         
+            func_8015CC28();
+            break;
+        }
+    }
+    g_Player.padHeld = g_Player.padPressed;
+    if (g_Player.D_80072EFC != 0) {
+        g_Player.D_80072EFC--;
+        g_Player.padPressed = g_Player.D_80072EF4;
+    } else {
+        g_Player.padPressed = g_pads[0].pressed;
+    }
+    g_Player.padTapped = (g_Player.padHeld ^ g_Player.padPressed) & g_Player.padPressed;
+    if (PLAYER.step == 0x10) {
+        goto block_47;
+    }
+    // Reuse the i variable here even though we aren't iterating
+    i = GetTeleportToOtherCastle();
+    if (i != TELEPORT_CHECK_NONE) {
+        func_8015CC70(i);
+    }
+    if (PLAYER.step == 0x20) {
+        goto block_48;
+    }
+    if ((g_DebugPlayer != 0) && (func_8015885C())) {
+        return;
+    }
+    if (g_Player.unk60 >= 2) {
+        goto block_47;
+    }
+    if (g_Player.unk60 == 1) {
+        var_s5 = PLAYER.step;
+        var_s6 = PLAYER.step_s;
+        SetPlayerStep(Player_BossGrab);
+        goto block_48;
+    }
+    if (((g_Player.D_80072F00[13] | g_Player.D_80072F00[14]) != 0)|| (PLAYER.unk44 == 0)) {
+        goto block_47;
+    }
+    var_s5 = PLAYER.step;
+    var_s6 = PLAYER.step_s;
+    sp10.effects = PLAYER.unk44 & 0xFFE0;
+    sp10.damageKind = PLAYER.unk44 & 0x1F;
+    sp10.damageTaken = PLAYER.hitPoints;
+    condition = g_api.func_800FD5BC(&sp10);
+    var_fp = sp10.damageKind;
+    var_s7 = sp10.effects;
+    if (condition) {
+        if (g_Player.unk5C == 0) {
+            SetPlayerStep(Player_Kill);
+        } else {
+            g_Status.hp = 1;
+            SetPlayerStep(Player_Hit);
+        }
+    } else {
+        SetPlayerStep(Player_Hit);
+    }
+    goto block_48;
+block_47:
+    CheckBladeDashInput();
+    CheckHighJumpInput();
+block_48:
+    g_Player.unk50 = PLAYER.step;
+    g_Player.unk52 = PLAYER.step_s;
+    switch (PLAYER.step) {                      
+    case 0:                                     
+        func_80158BFC();
+        break;
+    case 1:                                     
+        func_80158F38();
+        break;
+    case 2:                                     
+        func_80159670();
+        break;
+    case 3:                                     
+        func_801595D8();
+        break;
+    case 4:                                     
+        func_801590A0();
+        break;
+    case 8:                                     
+        func_8015C2A8();
+        break;
+    case 10:                                    
+        RichterHandleDamage(var_s7, var_fp, var_s5, var_s6);
+        break;
+    case 12:                                    
+        func_8015A7D0();
+        break;
+    case 16:                                    
+        func_8015A9B0(var_s7, var_fp, var_s5, var_s6);
+        break;
+    case 18:                                    
+        func_8015AFE0();
+        break;
+    case 19:                                    
+        func_8015B098();
+        break;
+    case 20:                                    
+        func_8015B18C();
+        break;
+    case 21:                                    
+        func_8015B244();
+        break;
+    case 27:                                    
+        func_8015B1E8();
+        break;
+    case 22:                                    
+        func_8015B348();
+        break;
+    case 23:                                    
+        func_8015B898();
+        break;
+    case 25:                                    
+        func_80158FA4();
+        break;
+    case 26:                                    
+        func_8015BE84();
+        break;
+    case 24:                                    
+        func_8015C178();
+        break;
+    case 32:                                    
+        func_8015BCD0();
+        break;
+    }
+    g_Player.unk08 = g_Player.unk0C;
+    switch (PLAYER.step) {                      /* switch 4 */
+    case 0:                                     /* switch 4 */
+    case 1:                                     /* switch 4 */
+        var_s4 = 0x08000000;
+        break;
+    case 2:                                     /* switch 4 */
+        var_s4 = 0x08000000;
+        if (PLAYER.step_s != 2) {
+            var_s4 = 0x08000020;
+        }
+        break;
+    case 3:                                     /* switch 4 */
+    case 4:                                     /* switch 4 */
+        var_s4 = 0x08002000;
+        break;
+    case 8:                                     /* switch 4 */
+        func_8015CAD4(1, 4);
+        break;
+    case 10:                                    /* switch 4 */
+        var_s4 = 0x08010000;
+    case 18:                                    /* switch 4 */
+        func_8015CAD4(1, 16);
+        break;
+    case 12:                                    /* switch 4 */
+        var_s4 = 0x08110040;
+        func_8015CAD4(1, 16);
+        break;
+    case 16:                                    /* switch 4 */
+        var_s4 = 0x08050000;
+        if (PLAYER.step_s == 0x80) {
+            var_s4 = 0x080D0000;
+        }
+        func_8015CAD4(1, 16);
+        break;
+    case 23:                                    /* switch 4 */
+    case 26:                                    /* switch 4 */
+        var_s4 = 0x20;
+        break;
+    case 19:                                    /* switch 4 */
+    case 20:                                    /* switch 4 */
+    case 21:                                    /* switch 4 */
+    case 22:                                    /* switch 4 */
+    case 27:                                    /* switch 4 */
+    case 32:                                    /* switch 4 */
+        var_s4 = 0x08000000;
+        func_8015CAD4(1, 16);
+        break;
+    }
+    if (g_Player.D_80072F00[9] != 0) {
+        var_s4 |= 0x400;
+    }
+    if (g_Player.D_80072F00[10] != 0) {
+        var_s4 |= 0x800;
+    }
+    if (g_Player.D_80072F00[12] != 0) {
+        var_s4 |= 0x1000;
+    }
+    if (*D_80097448 != 0) {
+        var_s4 |= 0x20000;
+    }
+    var_s4 |= 0x10000000;
+    g_Player.unk0C = var_s4;
+    if (g_Player.unk08 & 0x10000) {
+        if (!(var_s4 & 0x10000)) {
+            if (g_Player.unk5C != 0) {
+                if (g_Status.hp < 2) {
+                    func_8015D9B4();
+                    func_8015CAD4(1, 0x10);
+                }
+            } else {
+                func_8015CAD4(1, 0x10);
+                g_Player.D_80072F00[4] = 0x10;
+                PLAYER.palette = 0x8120;
+            }
+        }
+    }
+    if (var_s4 & 0x08000000) {
+        func_8015CB58(1, 4);
+    }
+    if ((g_Player.D_80072F00[13] | g_Player.D_80072F00[14]) != 0) {
+        g_Player.unk0C |= 0x100;
+    }
+    g_api.UpdateAnim(&D_80155964, &D_8015538C);
+    PLAYER.hitboxState = 1;
+    PLAYER.unk44 = 0;
+    PLAYER.hitPoints = 0;
+    g_Player.unk7A = 0;
+    if (PLAYER.unk4C == D_801556C4) {
+        PLAYER.palette = D_80154574[PLAYER.animFrameIdx];
+    }
+    if ((PLAYER.unk4C == D_80155658) && (PLAYER.animFrameIdx == 4)) {
+        PLAYER.palette = D_80154594[PLAYER.animFrameDuration & 3];
+    }
+    if ((PLAYER.step == 0x10) && (PLAYER.animFrameDuration < 0)) {
+        PLAYER.animCurFrame |= 0x8000;
+    }
+    if ((g_Player.unk0C & 0x50)) {
+        return;
+    }
+    func_8015C4AC();
+    if ((*D_80097448 >= 0x29) && (g_CurrentEntity->nFramesInvincibility == 0)) {
+        PLAYER.velocityY = PLAYER.velocityY * 3/4;
+        PLAYER.velocityX = PLAYER.velocityX * 3/4;
+    }
+    playerY = &g_Entities[PLAYER_CHARACTER].posY.i;
+    temp_s0 = g_Player.pl_vram_flag;
+    if ((ABS(PLAYER.velocityY) > FIX(2)) || (ABS(PLAYER.velocityX) > FIX(2))) {
+        PLAYER.velocityY = PLAYER.velocityY >> 2;
+        PLAYER.velocityX = PLAYER.velocityX >> 2;
+        if ((playerY->i.hi < 0) || 
+            (func_801572A8(1), (playerY->i.hi < 0)) || 
+            (func_801572A8(0), (playerY->i.hi < 0)) || 
+            (func_801572A8(0), (playerY->i.hi < 0)) || 
+            (func_801572A8(0), (playerY->i.hi < 0))) {
+            PLAYER.posY.val = FIX(-1);
+        }
+        PLAYER.velocityX *= 4;
+        PLAYER.velocityY *= 4;
+    } else {
+        func_801572A8(1);
+    }
+    g_Player.unk04 = temp_s0;
+    if ((*D_80097448 >= 0x29) && (g_CurrentEntity->nFramesInvincibility == 0)) {
+        PLAYER.velocityY = (PLAYER.velocityY * 4) / 3;
+        PLAYER.velocityX = (PLAYER.velocityX * 4) / 3;
+    }
+    g_CurrentEntity->nFramesInvincibility = 0;
+    func_8015C6D4();
+}
 
 void func_801587C0() { D_80175956 = 0; }
 
@@ -439,939 +778,4 @@ bool func_8015885C(void) {
     }
     FntPrint("charal:%02x\n", PLAYER.animCurFrame);
     return true;
-}
-
-void func_80158B04(s32 arg0) {
-    s32 var_s0;
-
-    if (PLAYER.facingLeft != 0) {
-        var_s0 = -3;
-    } else {
-        var_s0 = 3;
-    }
-
-    PLAYER.posY.i.hi -= 16;
-    PLAYER.posX.i.hi = var_s0 + PLAYER.posX.i.hi;
-    CreateEntFactoryFromEntity(g_CurrentEntity, FACTORY(0x100, 4), 0);
-    PLAYER.posY.i.hi += 16;
-    PLAYER.posX.i.hi = PLAYER.posX.i.hi - var_s0;
-
-    if (arg0 & 1) {
-        g_api.func_80102CD8(3);
-        g_api.PlaySfx(NA_SE_SECRET_STAIRS);
-    }
-
-    if (arg0 & 2) {
-        PLAYER.velocityX = 0;
-        PLAYER.velocityY = 0;
-    }
-}
-
-void func_80158BFC(void) {
-    s32 var_s0;
-
-    if (PLAYER.step < 64) {
-        if (D_8015459C != 0) {
-            D_8015459C--;
-        } else if (D_80097448[0] >= 49) {
-            if (PLAYER.facingLeft != 0) {
-                var_s0 = -4;
-            } else {
-                var_s0 = 4;
-            }
-            PLAYER.posX.i.hi = var_s0 + PLAYER.posX.i.hi;
-            PLAYER.posY.i.hi -= 16;
-            CreateEntFactoryFromEntity(g_CurrentEntity, FACTORY(0x800, 4), 0);
-            D_8015459C = 0x60;
-            PLAYER.posY.i.hi += 16;
-            PLAYER.posX.i.hi = PLAYER.posX.i.hi - var_s0;
-        }
-    }
-
-    if (func_8015DBB0(0x40000 | 0x305C) == false) {
-        func_8015C93C(0x2000);
-        switch (PLAYER.step_s) {
-        case 0:
-            if (func_8015C9CC() == 0) {
-                if (g_Player.padPressed & PAD_UP) {
-                    func_8015C920(&D_80155394);
-                    PLAYER.step_s = 1;
-                    break;
-                }
-            } else {
-                func_8015CDE0(0);
-            }
-            break;
-
-        case 1:
-            if (func_8015C9CC() != 0) {
-                func_8015CDE0(0);
-                break;
-            } else if (g_Player.padPressed & PAD_UP) {
-                break;
-            } else {
-                func_8015CD98(0);
-                break;
-            }
-
-        case 64:
-            func_8015CB58(1, 1);
-            if (PLAYER.animFrameIdx < 3) {
-                func_8015C9CC();
-                if (g_Player.padPressed & PAD_DOWN) {
-                    PLAYER.step = 2;
-                    PLAYER.unk4C = D_801555A8;
-                    break;
-                }
-            }
-
-            if (PLAYER.animFrameDuration < 0) {
-                if (g_Player.padPressed & PAD_SQUARE) {
-                    g_Player.unk46 = 2;
-                    PLAYER.step_s++;
-                    func_8015C920(D_80155730);
-                    CreateEntFactoryFromEntity(
-                        g_CurrentEntity, FACTORY(0, 17), 0);
-                    break;
-                }
-                g_Player.unk46 = 0;
-                func_8015CD98(0);
-            }
-            break;
-
-        case 65:
-            func_8015CB58(1, 1);
-            if (g_Player.padPressed & PAD_SQUARE) {
-                break;
-            }
-            g_Player.unk46 = 0;
-            func_8015CD98(0);
-            break;
-
-        case 66:
-            func_8015CB58(1, 1);
-            if (PLAYER.animFrameIdx < 3) {
-                func_8015C9CC();
-            }
-            if (PLAYER.animFrameDuration < 0) {
-                g_Player.unk46 = 0;
-                func_8015CD98(0);
-            }
-        }
-    }
-}
-
-void func_80158F38(void) {
-    if (func_8015DBB0(0x305C) == 0) {
-        func_8015C93C(0x2000);
-        if (func_8015C9CC() == 0) {
-            func_8015CD98(0);
-            return;
-        }
-
-        if (PLAYER.step_s == 0) {
-            SetSpeedX(0x14000);
-        }
-    }
-}
-
-void func_80158FA4(void) {
-    if (g_Player.unk7A != 0) {
-        func_8015CDE0(0);
-        return;
-    }
-    g_Player.D_80072F00[8] = 8;
-    g_Player.D_80072F00[1] = 8;
-
-    if (!func_8015DBB0(0x305C)) {
-        func_8015C93C(0x2000);
-        if (func_8015C9CC() == 0) {
-            func_8015CD98(0);
-            if (g_Player.D_80072F00[11] == 0) {
-                if (!(g_Player.pl_vram_flag & 0xC)) {
-                    func_8015C920(&D_8015539C);
-                    CreateEntFactoryFromEntity(
-                        g_CurrentEntity, FACTORY(0, 0), 0);
-                }
-            } else {
-                PLAYER.velocityX = 0;
-            }
-        } else if (PLAYER.step_s == 0) {
-            SetSpeedX(0x24000);
-        }
-    }
-}
-
-void func_801590A0(void) {
-    if ((D_80154570 == 0) && (PLAYER.velocityY < FIX(-1)) &&
-        !(g_Player.unk44 & 0x40) && !(g_Player.padPressed & PAD_CROSS)) {
-        PLAYER.velocityY = FIX(-1);
-    }
-    if ((g_Player.pl_vram_flag & 2) && (PLAYER.velocityY < FIX(-1))) {
-        PLAYER.velocityY = FIX(-0.25);
-        g_Player.unk44 |= 0x20;
-    }
-    if (func_8015DBB0(0x11009)) {
-        return;
-    }
-    switch (PLAYER.step_s) {
-    // Need at least one fake case to make the switch match
-    case 1:
-        return;
-    case 0:
-        func_8015C93C(0x1000);
-        if (func_8015C9CC()) {
-            if (g_Player.unk44 & 0x10) {
-                SetSpeedX(FIX(2.25));
-            } else {
-                SetSpeedX(FIX(1.25));
-            }
-            g_Player.unk44 &= ~4;
-        } else {
-            g_Player.unk44 &= ~0x10;
-            if ((PLAYER.animFrameIdx < 2) &&
-                ((LOW(g_Player.unk44) & 0xC) == 4) &&
-                (g_Player.padTapped & PAD_CROSS)) {
-                func_8015C920(D_8015555C);
-                SetSpeedX(FIX(-1.5));
-                PLAYER.velocityY = FIX(-2.625);
-                if (g_Player.unk72 != 0) {
-                    PLAYER.velocityY = 0;
-                }
-                PLAYER.step_s = 2;
-                g_Player.unk44 |= 0xA;
-                g_Player.unk44 &= ~4;
-            }
-        }
-        return;
-    case 0x40:
-        func_8015CB58(1, 1);
-        if (PLAYER.animFrameIdx < 3) {
-            if (func_8015C9CC() != 0) {
-                if (g_Player.unk44 & 0x10) {
-                    SetSpeedX(FIX(2.25));
-                } else {
-                    SetSpeedX(FIX(1.25));
-                }
-                g_Player.unk44 &= ~4;
-            } else {
-                g_Player.unk44 &= ~0x10;
-            }
-        } else {
-            if (((g_Player.padPressed & PAD_RIGHT) && !PLAYER.facingLeft) ||
-                ((g_Player.padPressed & PAD_LEFT) && PLAYER.facingLeft)) {
-                if (g_Player.unk44 & 0x10) {
-                    SetSpeedX(FIX(2.25));
-                } else {
-                    SetSpeedX(FIX(1.25));
-                }
-                SetSpeedX(FIX(1.25));
-                g_Player.unk44 &= ~4;
-            } else {
-                g_Player.unk44 &= ~0x10;
-            }
-        }
-        if (PLAYER.animFrameDuration < 0) {
-            if (g_Player.padPressed & PAD_SQUARE) {
-                g_Player.unk46 = 2;
-                PLAYER.step_s += 1;
-                func_8015C920(D_80155740);
-                CreateEntFactoryFromEntity(g_CurrentEntity, FACTORY(0, 17), 0);
-            } else {
-                g_Player.unk46 = 0;
-                PLAYER.step_s = 0;
-                func_8015C920(D_80155528);
-            }
-        }
-        return;
-    case 0x41:
-        func_8015CB58(1, 1);
-        if (!(g_Player.padPressed & PAD_SQUARE)) {
-            g_Player.unk46 = 0;
-            PLAYER.step_s = 0;
-            func_8015C920(D_80155528);
-        }
-        return;
-    case 0x42:
-        func_8015CB58(1, 1);
-        if (PLAYER.animFrameIdx < 3) {
-            if (func_8015C9CC() != 0) {
-                if (g_Player.unk44 & 0x10) {
-                    SetSpeedX(FIX(2.25));
-                } else {
-                    SetSpeedX(FIX(1.25));
-                }
-                g_Player.unk44 &= ~4;
-            } else {
-                g_Player.unk44 &= ~0x10;
-            }
-        } else {
-            if (((g_Player.padPressed & PAD_RIGHT) && !PLAYER.facingLeft) ||
-                ((g_Player.padPressed & PAD_LEFT) && PLAYER.facingLeft)) {
-                if (g_Player.unk44 & 0x10) {
-                    SetSpeedX(FIX(2.25));
-                } else {
-                    SetSpeedX(FIX(1.25));
-                }
-                SetSpeedX(FIX(1.25));
-                g_Player.unk44 &= ~4;
-            } else {
-                g_Player.unk44 &= ~0x10;
-            }
-        }
-        if (PLAYER.animFrameDuration < 0) {
-            g_Player.unk46 = 0;
-            PLAYER.step_s = 0;
-            func_8015C920(D_80155528);
-        }
-    }
-}
-
-void func_801595D8(void) {
-    if (func_8015DBB0(0x9009)) {
-        return;
-    }
-    func_8015C93C(0x1000);
-    if (PLAYER.step_s != 0) {
-        return;
-    }
-    if (g_Player.D_80072F00[5] != 0 && g_Player.padTapped & PAD_CROSS) {
-        func_8015D020();
-    } else if (func_8015C9CC() != 0) {
-        SetSpeedX(0xC000);
-    }
-}
-
-void func_80159670(void) {
-    s32 i;
-    s16 xShift;
-
-    if ((g_Player.padTapped & PAD_CROSS) && (g_Player.unk46 == 0) &&
-        (g_Player.padPressed & PAD_DOWN)) {
-        for (i = 0; i < 4; i++) {
-            if ((g_Player.colliders[i].effects & EFFECT_SOLID_FROM_ABOVE)) {
-                g_Player.D_80072F00[7] = 8;
-                return;
-            }
-        }
-    }
-    if (func_8015DBB0(0x4100C)) {
-        return;
-    }
-    if ((g_Player.padTapped & PAD_CROSS) && (g_Player.unk46 == 0) &&
-        (g_Player.unk72 == 0)) {
-        func_8015D020(1);
-        return;
-    } else if ((g_Player.unk72 == 0) && (g_Player.unk46 == 0) &&
-               (g_Player.padTapped & PAD_TRIANGLE) && func_8015D678()) {
-        return;
-    }
-    func_8015C93C(0x2000);
-    switch (PLAYER.step_s) {
-    case 0x0:
-        if (D_8015459C != 0) {
-            D_8015459C--;
-        } else if ((*D_80097448 >= 0x19) && (g_Player.unk48 == 0)) {
-            if (PLAYER.facingLeft) {
-                xShift = -9;
-            } else {
-                xShift = 9;
-            }
-            PLAYER.posX.i.hi += xShift;
-            PLAYER.posY.i.hi += 2;
-            CreateEntFactoryFromEntity(g_CurrentEntity, FACTORY(0x800, 4), 0);
-            D_8015459C = 0x60;
-            PLAYER.posY.i.hi -= 2;
-            PLAYER.posX.i.hi -= xShift;
-        }
-        if (!(g_Player.padPressed & PAD_DOWN) &&
-            ((g_Player.unk72 == 0) || !(g_Player.pl_vram_flag & 0x40))) {
-            func_8015C920(D_801554E0);
-            PLAYER.step_s = 2;
-            return;
-        }
-        break;
-    case 0x1:
-        if (!(g_Player.padPressed & PAD_DOWN) &&
-            ((g_Player.unk72 == 0) || !(g_Player.pl_vram_flag & 0x40))) {
-            if (func_8015C9CC() == 0) {
-                PLAYER.unk4C = D_801554E0;
-                PLAYER.step_s = 2;
-                PLAYER.animFrameDuration = 1;
-                PLAYER.animFrameIdx = 2 - PLAYER.animFrameIdx;
-                return;
-            }
-            func_8015CDE0(0);
-            return;
-        }
-    case 0x4:
-        if (PLAYER.animFrameDuration != -1) {
-            return;
-        }
-        func_8015C920(D_801554C0);
-        PLAYER.step_s = 0;
-        return;
-    case 0x2:
-        if ((g_Player.unk72 == 0) || !(g_Player.pl_vram_flag & 0x40)) {
-            if (func_8015C9CC() != 0) {
-                func_8015CDE0(0);
-                return;
-            }
-            if (PLAYER.animFrameDuration == -1) {
-                func_8015CD98(0);
-                return;
-            }
-        }
-        break;
-    case 0x3:
-        if (PLAYER.animFrameDuration < 0) {
-            func_8015C920(D_801554C0);
-            PLAYER.step_s = 0;
-            return;
-        }
-        break;
-    case 0x40:
-        func_8015CB58(1, 1);
-        if (PLAYER.animFrameIdx < 3) {
-            func_8015C9CC();
-            if (!(g_Player.padPressed & PAD_DOWN) && (g_Player.unk72 == 0)) {
-                PLAYER.step = 0;
-                PLAYER.unk4C = D_80155588;
-                return;
-            }
-        }
-        if (PLAYER.animFrameDuration < 0) {
-            if (g_Player.padPressed & PAD_SQUARE) {
-                g_Player.unk46 = 2;
-                PLAYER.step_s++;
-                func_8015C920(D_80155738);
-                CreateEntFactoryFromEntity(g_CurrentEntity, FACTORY(0, 17), 0);
-                return;
-            }
-            g_Player.unk46 = 0;
-            PLAYER.step_s = 0;
-            func_8015C920(D_801554C0);
-        }
-        break;
-    case 0x41:
-        func_8015CB58(1, 1);
-        if (!(g_Player.padPressed & PAD_SQUARE)) {
-            g_Player.unk46 = 0;
-            PLAYER.step_s = 0;
-            func_8015C920(D_801554C0);
-        }
-        break;
-    }
-}
-
-void func_80159BC8(void) {
-    PLAYER.animFrameDuration = 0;
-    PLAYER.animFrameIdx = 0;
-    g_Player.unk44 = 0;
-    g_Player.unk46 = 0;
-    PLAYER.drawFlags &= 0xFB;
-}
-
-void func_80159C04(void) {
-    Entity* entity = PLAYER.ext.generic.unkB8.entityPtr;
-    s16 temp_v0;
-    s32 var_a0;
-    s32 var_a2;
-
-    if (entity->facingLeft != 0) {
-        var_a2 = -entity->hitboxOffX;
-    } else {
-        var_a2 = entity->hitboxOffX;
-    }
-
-    if (PLAYER.facingLeft != 0) {
-        var_a0 = -PLAYER.hitboxOffX;
-    } else {
-        var_a0 = PLAYER.hitboxOffX;
-    }
-
-    temp_v0 = var_a0 + PLAYER.posX.i.hi - entity->posX.i.hi - var_a2;
-
-    if (ABS(temp_v0) < 16) {
-        if (entity->velocityX != 0) {
-            if (entity->velocityX < 0) {
-                PLAYER.entityRoomIndex = 0;
-                return;
-            } else {
-                PLAYER.entityRoomIndex = 1;
-                return;
-            }
-        }
-    }
-
-    if (temp_v0 < 0) {
-        PLAYER.entityRoomIndex = 0;
-    } else {
-        PLAYER.entityRoomIndex = 1;
-    }
-}
-
-void RichterHandleDamage(s32 arg0, u32 arg1, s16 arg2) {
-    DamageParam damage;
-    s32 xShift;
-    s32 i;
-    bool step_s_zero = false;
-
-    if (D_80173B64 != 0) {
-        D_80173B64--;
-    }
-    switch (PLAYER.step_s) {
-    case 0:
-        step_s_zero = true;
-        func_80159BC8();
-        if (arg1 < 16) {
-            func_80159C04();
-        } else {
-            PLAYER.entityRoomIndex = PLAYER.facingLeft;
-        }
-        if (arg0 & 0x4000) {
-            CreateEntFactoryFromEntity(g_CurrentEntity, FACTORY(0, 48), 0);
-            PLAYER.velocityY = FIX(-4);
-            func_8015CAAC(FIX(-1.25));
-            PLAYER.step_s = 1;
-            PLAYER.unk4C = D_801556C4;
-            g_Player.unk40 = 0x8120;
-            CreateEntFactoryFromEntity(g_CurrentEntity, FACTORY(0, 48), 0);
-            CreateEntFactoryFromEntity(g_CurrentEntity, FACTORY(0x4600, 33), 0);
-            g_Player.D_80072F00[2] = 6;
-            g_api.PlaySfx(0x703);
-            break;
-        } else {
-            if (arg0 & 0x2000) {
-                arg1 = 3;
-            }
-            switch (arg1 - 2) {
-            case 0:
-                switch (arg2) {
-                case 0:
-                case 1:
-                    PLAYER.velocityY = 0;
-                    func_8015CAAC(FIX(-1.25));
-                    PLAYER.step_s = 6;
-                    PLAYER.unk4C = D_8015569C;
-                    g_api.PlaySfx(0x702);
-                    CreateEntFactoryFromEntity(
-                        g_CurrentEntity, FACTORY(0, 0), 0);
-                    break;
-                case 2:
-                    PLAYER.velocityY = 0;
-                    func_8015CAAC(FIX(-1.25));
-                    PLAYER.step_s = 7;
-                    PLAYER.unk4C = D_80155704;
-                    CreateEntFactoryFromEntity(
-                        g_CurrentEntity, FACTORY(0, 0), 0);
-                    g_api.PlaySfx(0x703);
-                    break;
-                case 3:
-                case 4:
-                    PLAYER.velocityY = FIX(-3);
-                    func_8015CAAC(FIX(-1.25));
-                    PLAYER.step_s = 1;
-                    PLAYER.unk4C = D_80155694;
-                    g_api.PlaySfx(0x702);
-                    break;
-                }
-                break;
-            case 2:
-            case 3:
-                PLAYER.velocityY = FIX(-0.5);
-                g_Player.damageTaken = PLAYER.hitPoints;
-                PLAYER.posY.val -= 1;
-                func_8015CAAC(FIX(-8));
-                PLAYER.step_s = 2;
-                PLAYER.unk4C = D_80155694;
-                g_Player.D_80072F00[2] = 0x200;
-                PLAYER.facingLeft = PLAYER.entityRoomIndex;
-                break;
-            default:
-            case 1:
-            case 4:
-            case 5:
-            case 6:
-            case 7:
-            case 8:
-            case 9:
-            case 10:
-            case 11:
-            case 12:
-            case 13:
-            case 14:
-                switch (arg2) {
-                default:
-                case 0:
-                case 1:
-                    PLAYER.velocityY = FIX(-4);
-                    func_8015CAAC(FIX(-1.25));
-                    PLAYER.step_s = 1;
-                    PLAYER.unk4C = D_80155694;
-                    g_api.PlaySfx(0x702);
-                    break;
-                case 2:
-                    PLAYER.velocityY = 0;
-                    func_8015CAAC(FIX(-1.25));
-                    PLAYER.step_s = 7;
-                    PLAYER.unk4C = D_80155704;
-                    CreateEntFactoryFromEntity(
-                        g_CurrentEntity, FACTORY(0, 0), 0);
-                    g_api.PlaySfx(0x703);
-                    break;
-                case 3:
-                case 4:
-                    PLAYER.velocityY = FIX(-3);
-                    func_8015CAAC(FIX(-1.25));
-                    PLAYER.step_s = 1;
-                    PLAYER.unk4C = D_80155694;
-                    g_api.PlaySfx(0x702);
-                    break;
-                }
-                break;
-            }
-            g_Player.unk40 = 0x8166;
-            g_Player.D_80072F00[2] = 6;
-            if (arg0 & 0x8000) {
-                CreateEntFactoryFromEntity(
-                    g_CurrentEntity, FACTORY(0x100, 8), 0);
-                CreateEntFactoryFromEntity(g_CurrentEntity, FACTORY(0, 9), 0);
-                CreateEntFactoryFromEntity(
-                    g_CurrentEntity, FACTORY(0x4300, 33), 0);
-                g_Player.unk40 = 0x8160;
-                g_Player.D_80072F00[2] = 0x10;
-                break;
-            } else if (arg0 & 0x40) {
-                CreateEntFactoryFromEntity(
-                    g_CurrentEntity, FACTORY(0x500, 46), 0);
-                CreateEntFactoryFromEntity(
-                    g_CurrentEntity, FACTORY(0x4400, 33), 0);
-                g_Player.unk40 = 0x8166;
-                g_Player.D_80072F00[2] = 0x10;
-                break;
-            } else if (arg0 & 0x2000) {
-                // Creates EntityHitByIce
-                CreateEntFactoryFromEntity(g_CurrentEntity, FACTORY(0, 47), 0);
-                g_Player.D_80072F00[2] = 0xC;
-                g_Player.unk40 = 0x8169;
-                break;
-            } else {
-                if (arg0 & 0x800) {
-                    CreateEntFactoryFromEntity(
-                        g_CurrentEntity, FACTORY(0, 70), 0);
-                    CreateEntFactoryFromEntity(
-                        g_CurrentEntity, FACTORY(0x5600, 33), 0);
-                    g_Player.D_80072F00[2] = 0x10;
-                    g_Player.unk40 = 0x8164;
-                }
-                if (arg0 & 0x1000) {
-                    CreateEntFactoryFromEntity(
-                        g_CurrentEntity, FACTORY(0, 71), 0);
-                    CreateEntFactoryFromEntity(
-                        g_CurrentEntity, FACTORY(0x5700, 33), 0);
-                    g_Player.D_80072F00[2] = 8;
-                    g_Player.unk40 = 0x8168;
-                }
-                if (!(arg0 & 0xF840)) {
-                    CreateEntFactoryFromEntity(
-                        g_CurrentEntity, FACTORY(0x5300, 33), 0);
-                }
-            }
-        }
-        break;
-    case 1:
-        if ((g_Player.pl_vram_flag & 2) && (PLAYER.velocityY < FIX(-1))) {
-            PLAYER.velocityY = FIX(-1);
-        }
-        if (func_8015DBB0(0x20280)) {
-            return;
-        }
-        break;
-    case 2:
-        if ((g_Player.unk04 & 0x8000) && !(g_Player.pl_vram_flag & 0x8000)) {
-            goto block_6dc;
-        }
-        if ((g_Player.pl_vram_flag & 0x8000) && !(g_GameTimer & 1)) {
-            CreateEntFactoryFromEntity(g_CurrentEntity, FACTORY(0xA00, 74), 0);
-        }
-        if (!(g_Player.pl_vram_flag & 0xE)) {
-            break;
-        }
-        if (g_Player.pl_vram_flag & 2) {
-            func_80158B04(1);
-            D_80173B64 = 0x18;
-            PLAYER.velocityX /= 2;
-            PLAYER.velocityY = 0;
-            PLAYER.step_s = 5;
-            damage.effects = 0;
-            damage.damageKind = 1;
-            damage.damageTaken = g_Player.damageTaken;
-            if (g_api.func_800FD5BC(&damage) != 0) {
-                SetPlayerStep(Player_Kill);
-                func_8015A9B0(0, 2, 10, 2);
-                return;
-            }
-            break;
-        } else {
-            if ((g_StageId != STAGE_BO6) && (g_StageId != STAGE_RBO6) &&
-                (g_StageId != STAGE_DRE)) {
-                for (i = 2; i < 7; i++) {
-                    if ((g_Player.colliders3[i].effects & 2)) {
-                        break;
-                    }
-                }
-                if (i == 7) {
-                    for (i = 9; i < 0xE; i++) {
-                        if ((g_Player.colliders3[i].effects & 2)) {
-                            break;
-                        }
-                    }
-                }
-                if (i == 14) {
-                block_6dc:
-                    PLAYER.velocityY = FIX(-4);
-                    func_8015CAAC(FIX(-1.25));
-                    xShift = -3;
-                    if (PLAYER.velocityX != 0) {
-                        xShift = 3;
-                    }
-                    PLAYER.posY.i.hi += 20;
-                    PLAYER.posX.i.hi = xShift + PLAYER.posX.i.hi;
-                    CreateEntFactoryFromEntity(
-                        g_CurrentEntity, FACTORY(0x900, 4), 0);
-                    PLAYER.posY.i.hi -= 20;
-                    PLAYER.posX.i.hi -= xShift;
-                    g_api.PlaySfx(NA_SE_EN_ROCK_BREAK);
-                    g_api.func_80102CD8(2);
-                    PLAYER.step_s = 1;
-                    damage.effects = 0;
-                    damage.damageKind = 1;
-                    damage.damageTaken = g_Player.damageTaken;
-                    if (g_api.func_800FD5BC(&damage) != 0) {
-                        SetPlayerStep(Player_Kill);
-                        func_8015A9B0(0, 2, 10, 2);
-                        return;
-                    }
-                    break;
-                }
-            }
-            D_80173B64 = 8;
-            g_api.PlaySfx(NA_SE_EN_ROCK_BREAK);
-            PLAYER.velocityY = FIX(-2.5);
-            g_api.func_80102CD8(2);
-            PLAYER.step_s = 3;
-            CreateEntFactoryFromEntity(g_CurrentEntity, FACTORY(0x800, 75), 0);
-        }
-        damage.effects = 0;
-        damage.damageKind = 1;
-        damage.damageTaken = g_Player.damageTaken;
-        if (g_api.func_800FD5BC(&damage) != 0) {
-            SetPlayerStep(Player_Kill);
-            func_8015A9B0(0, 2, 10, 2);
-            return;
-        }
-        break;
-    case 3:
-        if (D_80173B64 == 0) {
-            SetSpeedX(FIX(0.75));
-            if (func_8015DBB0(0x20280)) {
-                return;
-            }
-        }
-        break;
-    case 5:
-        func_8015C93C(0x2000);
-        if (D_80173B64 != 0) {
-            if ((g_Player.pl_vram_flag & 2) && !(g_GameTimer & 3)) {
-                func_80158B04(0);
-            }
-            break;
-        } else if (g_Player.pl_vram_flag & 0xC) {
-            if (!(g_Player.pl_vram_flag & 0xFF03)) {
-                PLAYER.velocityY += FIX(12.0 / 128);
-                if (PLAYER.velocityY > FIX(7)) {
-                    PLAYER.velocityY = FIX(7);
-                }
-                if (!(g_GameTimer & 3)) {
-                    CreateEntFactoryFromEntity(
-                        g_CurrentEntity, FACTORY(0x400, 76), 0);
-                }
-                break;
-            }
-        }
-        PLAYER.step_s = 1;
-        PLAYER.animFrameIdx = 0;
-        PLAYER.animFrameDuration = 0;
-        break;
-    case 6:
-        func_8015C93C(0x2000);
-        if (!(g_Player.pl_vram_flag & 1)) {
-            func_8015CF08();
-        }
-        if (PLAYER.animFrameDuration < 0) {
-            if ((g_Player.unk5C == 0) || (g_Status.hp > 0)) {
-                func_8015CD98(PLAYER.velocityX);
-                break;
-            }
-            func_8015D9B4();
-            return;
-        }
-        break;
-    case 7:
-        func_8015C93C(0x2000);
-        if (!(g_Player.pl_vram_flag & 1)) {
-            func_8015CF08();
-        }
-        if (PLAYER.animFrameDuration < 0) {
-            if ((g_Player.unk5C != 0) && (g_Status.hp <= 0)) {
-                func_8015D9B4();
-                return;
-            }
-            func_8015CCC8(0, PLAYER.velocityX);
-        }
-        break;
-    }
-    if (step_s_zero && (g_Player.unk72 != 0)) {
-        PLAYER.velocityY = 0;
-    }
-}
-
-const s32 rodata_padding_1A784 = 0;
-
-void func_8015A7D0(void) {
-    DamageParam damage;
-    switch (g_CurrentEntity->step_s) {
-    case 0:
-        func_80159BC8();
-        if (g_Player.unk62 == 0) {
-            PLAYER.unk4C = D_80155694;
-            g_Player.D_80072F00[2] = 8;
-            g_Player.unk40 = 0x8166;
-            g_api.PlaySfx(0x702);
-        }
-        PLAYER.step_s = 1;
-        PLAYER.velocityX = PLAYER.velocityY = 0;
-        g_Player.unk60 = 2;
-        return;
-    case 1:
-        // Effectively a switch on g_Player.unk60
-        if (g_Player.unk60 == 3) {
-            damage.effects = 0;
-            damage.damageKind = 1;
-            damage.damageTaken = g_Player.damageTaken;
-            if (g_api.func_800FD5BC(&damage)) {
-                SetPlayerStep(Player_Kill);
-                func_8015A9B0(0, 2, 12, 1);
-                return;
-            }
-            if (g_Player.unk62 == 0) {
-                g_Player.D_80072F00[2] = 4;
-                g_Player.unk40 = 0x8166;
-                g_api.PlaySfx(0x703);
-            }
-            g_Player.unk60 = 2;
-        } else if (g_Player.unk60 == 4) { /* switch 1 */
-            PLAYER.step = Player_Hit;
-            PLAYER.step_s = 2;
-            g_Player.unk60 = 0;
-            g_Player.damageTaken = g_Player.unk64;
-            g_api.PlaySfx(0x704);
-        } else if (g_Player.unk60 == 0) { /* switch 1 */
-            PLAYER.step = Player_Hit;
-            PLAYER.step_s = 1;
-        }
-    }
-}
-
-INCLUDE_ASM("ric/nonmatchings/1AC60", func_8015A9B0);
-
-void func_8015AFE0(void) {
-    if (PLAYER.step_s == 0) {
-        PLAYER.velocityY += 0x3800;
-        if (PLAYER.velocityY > 0) {
-            PLAYER.velocityY = 0;
-            PLAYER.step_s = 1;
-        }
-    } else if (g_Player.unk4E != 0) {
-        g_Player.unk46 = 0;
-        SetPlayerStep(4);
-        func_8015C920(&D_80155528);
-        g_Player.unk44 = 0;
-    }
-    if (g_Player.unk72 != 0) {
-        PLAYER.velocityY = 0;
-    }
-}
-
-void func_8015B098(void) {
-    if ((PLAYER.animCurFrame == 0xB5) && (PLAYER.animFrameDuration == 1)) {
-        CreateEntFactoryFromEntity(g_CurrentEntity, FACTORY(0, 35), 0);
-        g_api.PlaySfx(NA_SE_UNK_62F);
-    }
-
-    if (PLAYER.animFrameDuration < 0) {
-        func_8015CD98(0);
-        g_Player.unk46 = 0;
-        CreateEntFactoryFromEntity(g_CurrentEntity, FACTORY(0x4500, 33), 0);
-        g_Player.D_80072F00[0] = 0x800;
-    }
-
-    if (!(g_Player.pl_vram_flag & 1)) {
-        func_8015CF08();
-        g_Player.unk46 = 0;
-        CreateEntFactoryFromEntity(g_CurrentEntity, FACTORY(0x4500, 33), 0);
-        g_Player.D_80072F00[0] = 0x800;
-    }
-}
-
-void func_8015B18C(void) {
-    if (PLAYER.animFrameDuration < 0) {
-        func_8015CD98(0);
-        g_Player.unk46 = 0;
-    }
-
-    if ((g_Player.pl_vram_flag & 1) == 0) {
-        func_8015CF08();
-        g_Player.unk46 = 0;
-    }
-}
-
-void func_8015B1E8(void) {
-    if (g_Player.unk4E) {
-        func_8015CD98(0);
-        g_Player.unk46 = 0;
-    }
-
-    if ((g_Player.pl_vram_flag & 1) == 0) {
-        func_8015CF08();
-        g_Player.unk46 = 0;
-    }
-}
-
-void func_8015B244(void) {
-    if (PLAYER.step_s == 0) {
-        D_80174F74 = 0x200;
-        PLAYER.step_s++;
-    } else {
-        func_8015C9CC();
-        D_80174F74--;
-        if (D_80174F74 == 0) {
-            g_Player.unk46 = 0;
-            func_8015CD98(0);
-            g_Player.unk4E = 1;
-        }
-    }
-    if (g_Player.padTapped & PAD_CROSS) {
-        func_8015D020();
-        g_Player.unk46 = 0;
-        g_Player.unk4E = 1;
-        D_80174F74 = 0;
-    }
-    if (!(g_Player.pl_vram_flag & 1)) {
-        func_8015CF08();
-        g_Player.unk46 = 0;
-        g_Player.unk4E = 1;
-        D_80174F74 = 0;
-    }
 }
